Fixes startbit > 255 and table values with extended id's (#11)
* Fix for message with startbit > 255

* Table values with extended id are added to signals

+++ DbcParserLib/DbcParser.cs
@@ -39,7 +39,7 @@ namespace DbcParserLib
     {
         public uint ID;
         public string Name;
-        public byte StartBit;
+        public ushort StartBit;
         public byte Length;
         public byte ByteOrder = 1;
         public byte IsSigned;
